lightweight_real_esrgan_anime/requirements.txt

huggingface-hub
onnxruntime
Pillow
numpy